Address PR #525 review feedback: rename data source to PrefsDataStoreSettingsDataSource, remove Hilt dependencies from core:settings, add KDocs, and fix dependencies
