+++ app/src/main/java/com/google/jetpackcamera/AppSettingsModule.kt
@@ -13,35 +13,42 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.jetpackcamera.core.settings.datastoreprefs
+package com.google.jetpackcamera
 
 import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.PreferenceDataStoreFactory
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.preferencesDataStoreFile
+import com.google.jetpackcamera.core.common.DefaultCaptureModeOverride
+import com.google.jetpackcamera.core.settings.datastoreprefs.PrefsDataStoreSettingsDataSource
+import com.google.jetpackcamera.model.CaptureMode
+import com.google.jetpackcamera.settings.SettingsDataSource
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.SupervisorJob
 
 @Module
 @InstallIn(SingletonComponent::class)
-object DataStoreModule {
-    private const val FILE_LOCATION = "jca_app_settings.preferences_pb"
+object AppSettingsModule {
 
     @Provides
     @Singleton
-    fun providePreferenceDataStore(@ApplicationContext context: Context): DataStore<Preferences> =
-        PreferenceDataStoreFactory.create(
-            scope = CoroutineScope(Dispatchers.IO + SupervisorJob()),
-            produceFile = {
-                context.preferencesDataStoreFile(FILE_LOCATION)
-            }
+    fun providePreferencesDataStore(@ApplicationContext context: Context): DataStore<Preferences> {
+        return PreferenceDataStoreFactory.create(
+            produceFile = { context.preferencesDataStoreFile("app_settings.preferences_pb") }
         )
+    }
+
+    @Provides
+    @Singleton
+    fun provideSettingsDataSource(
+        dataStore: DataStore<Preferences>,
+        @DefaultCaptureModeOverride defaultCaptureMode: CaptureMode
+    ): SettingsDataSource {
+        return PrefsDataStoreSettingsDataSource(dataStore, defaultCaptureMode)
+    }
 }
