+++ feature/settings/src/androidTest/java/com/google/jetpackcamera/settings/CameraAppSettingsViewModelTest.kt
@@ -21,7 +21,7 @@ import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.stringPreferencesKey
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
-import com.google.jetpackcamera.core.settings.datastoreprefs.LocalSettingsDataSource
+import com.google.jetpackcamera.core.settings.datastoreprefs.PrefsDataStoreSettingsDataSource
 import com.google.jetpackcamera.core.settings.datastoreprefs.testing.FakeDataStoreModule
 import com.google.jetpackcamera.model.CaptureMode
 import com.google.jetpackcamera.model.DarkMode
@@ -68,7 +68,7 @@ internal class CameraAppSettingsViewModelTest {
             file = testFile
         )
 
-        val settingsDataSource = LocalSettingsDataSource(
+        val settingsDataSource = PrefsDataStoreSettingsDataSource(
             dataStore = testDataStore,
             defaultCaptureModeOverride = CaptureMode.STANDARD
         )
